# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Rocks Accounts API
  description: Mock Accounts API for study purposes
  version: 1.0.0

servers:
  - url: http://localhost:3003

paths:
  /v1/accounts:
    post:
      summary: Open a new account
      operationId: openAccount
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
          description: Required idempotency key. Same key+payload → 409; same key+different payload → 400.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpenAccountResponse'
        '400':
          description: Invalid request or mismatched idempotency payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate idempotency key with same payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/accounts/{accountId}:
    get:
      summary: Get account details
      operationId: getAccount
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    OpenAccountRequest:
      type: object
      required:
        - name
        - document
      properties:
        name:
          type: string
          minLength: 3
        document:
          type: string
          pattern: '^[0-9]{11}$'

    OpenAccountResponse:
      type: object
      required:
        - accountId
        - number
        - status
        - createdAt
        - name
      properties:
        accountId:
          type: string
          format: uuid
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        number:
          type: string
          pattern: '^[0-9]{6}$'
          example: '123456'
        branch:
          type: string
          pattern: '^[0-9]{4}$'
          example: '0001'
        name:
          type: string
          minLength: 3
          example: 'João da Silva'
        status:
          type: string
          enum: [active]
        createdAt:
          type: string
          format: date-time
          example: '2025-08-29T12:00:00Z'

    Account:
      allOf:
        - $ref: '#/components/schemas/OpenAccountResponse'
        - type: object
          required:
            - document
          properties:
            document:
              type: string
              pattern: '^[0-9]{11}$'
              example: '00000000000'
            productCode:
              type: string
              nullable: true
              example: 'CA_BASIC'
            currency:
              type: string
              nullable: true
              example: 'BRL'

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string